Use the resource version from the terraform state
For some resource the version from the state was used and for other not.
Make this consistent and use the state version where possible.

There are two exceptions, the tax and shipping_method resources. Reason
is that they both have 'fake' child resources
(tax_rate, shipping_method_rate) which modifies the version of the parent