Use a ListNestedBlock for project settings
The step to using a SingleNestedBlock failed due to some issues with terraform, described at https://github.com/hashicorp/terraform-plugin-framework/issues/603

Instead use a ListNestedBlock with a maximum size of 1. This basically mirrors the approach used in the terraform-plugin-sdk and seems to work stable. It also removes the need to migrate existing state from a list to a object.